Demote "Missing article on server" message from warning to debug. Don't try to remove workdir if it does not exist. Closes #149 for trunk.

+++ main/sabnzbd/postproc.py
@@ -196,7 +196,8 @@ class PostProcessor(Thread):
 
             ## Remove download folder
             try:
-                os.rmdir(workdir)
+                if os.path.exists(workdir):
+                    os.rmdir(workdir)
             except:
                 logging.error("[%s] Error removing workdir (%s)", __NAME__, workdir)
